requirements.txt
flask>=3.0.0
pywebview>=5.0.0 